add client and fakeclient files
