+++ pkg/docker/docker_test.go
@@ -12,205 +12,11 @@
 package docker
 
 import (
-	"bytes"
-	"context"
-	"errors"
-	"io"
-	"io/ioutil"
 	"testing"
-	"time"
 
 	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/api/types/container"
-	"github.com/docker/docker/api/types/registry"
-	v1 "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/stretchr/testify/assert"
 )
-
-var mockImageSummaryWithCwImages = []types.ImageSummary{
-	types.ImageSummary{
-		ID:          "pfe",
-		RepoDigests: []string{"eclipse/codewind-pfe", "sha256:7173b809", "test:0.0.9"},
-		RepoTags:    []string{"test:0.0.9"},
-	},
-	types.ImageSummary{
-		ID:          "performance",
-		RepoDigests: []string{"eclipse/codewind-performance", "sha256:7173b809", "test:0.0.9"},
-		RepoTags:    []string{"test:0.0.9"},
-	},
-}
-
-var mockContainerListWithCwContainers = []types.Container{
-	types.Container{
-		Names: []string{"/codewind-pfe"},
-		ID:    "pfe",
-		Image: "eclipse/codewind-pfe:0.0.9",
-		Ports: []types.Port{types.Port{PrivatePort: 9090, PublicPort: 1000, IP: "pfe"}}},
-	types.Container{
-		Names: []string{"/codewind-performance"},
-		Image: "eclipse/codewind-performance:0.0.9"},
-}
-
-var mockImageSummaryWithoutCwImages = []types.ImageSummary{
-	types.ImageSummary{
-		ID:       "golang",
-		RepoTags: []string{"golang:0.0.9"},
-	},
-	types.ImageSummary{
-		ID:       "registry",
-		RepoTags: []string{"registry:0.0.9"},
-	},
-}
-
-var mockContainerListWithoutCwContainers = []types.Container{
-	types.Container{
-		Names: []string{"/registry"},
-		Image: "registry",
-	},
-	types.Container{
-		Names: []string{"/go-test"},
-		Image: "golang",
-	},
-}
-
-// This mock client will return container and images lists, with Codewind items included
-type mockDockerClientWithCw struct {
-}
-
-func (m *mockDockerClientWithCw) ImagePull(ctx context.Context, image string, imagePullOptions types.ImagePullOptions) (io.ReadCloser, error) {
-	r := ioutil.NopCloser(bytes.NewReader([]byte("")))
-	return r, nil
-}
-
-func (m *mockDockerClientWithCw) ImageList(ctx context.Context, imageListOptions types.ImageListOptions) ([]types.ImageSummary, error) {
-	return mockImageSummaryWithCwImages, nil
-}
-
-func (m *mockDockerClientWithCw) ContainerList(ctx context.Context, containerListOptions types.ContainerListOptions) ([]types.Container, error) {
-	return mockContainerListWithCwContainers, nil
-}
-
-func (m *mockDockerClientWithCw) ContainerStop(ctx context.Context, containerID string, timeout *time.Duration) error {
-	return nil
-}
-
-func (m *mockDockerClientWithCw) ContainerRemove(ctx context.Context, containerID string, options types.ContainerRemoveOptions) error {
-	return nil
-}
-
-func (m *mockDockerClientWithCw) ContainerInspect(ctx context.Context, containerID string) (types.ContainerJSON, error) {
-	return types.ContainerJSON{
-		ContainerJSONBase: &types.ContainerJSONBase{
-			HostConfig: &container.HostConfig{
-				AutoRemove: true,
-			},
-		},
-	}, nil
-}
-
-func (m *mockDockerClientWithCw) DistributionInspect(ctx context.Context, image, encodedRegistryAuth string) (registry.DistributionInspect, error) {
-	return registry.DistributionInspect{
-		Descriptor: v1.Descriptor{
-			Digest: "sha256:7173b809",
-		},
-	}, nil
-}
-
-func (m *mockDockerClientWithCw) RegistryLogin(ctx context.Context, auth types.AuthConfig) (registry.AuthenticateOKBody, error) {
-	return registry.AuthenticateOKBody{}, nil
-}
-
-// This mock client will return valid image and containers lists, without Codewind items
-type mockDockerClientWithoutCw struct {
-}
-
-func (m *mockDockerClientWithoutCw) ImagePull(ctx context.Context, image string, imagePullOptions types.ImagePullOptions) (io.ReadCloser, error) {
-	r := ioutil.NopCloser(bytes.NewReader([]byte("")))
-	return r, nil
-}
-
-func (m *mockDockerClientWithoutCw) ImageList(ctx context.Context, imageListOptions types.ImageListOptions) ([]types.ImageSummary, error) {
-	return mockImageSummaryWithoutCwImages, nil
-}
-
-func (m *mockDockerClientWithoutCw) ContainerList(ctx context.Context, containerListOptions types.ContainerListOptions) ([]types.Container, error) {
-	return mockContainerListWithoutCwContainers, nil
-}
-
-func (m *mockDockerClientWithoutCw) ContainerStop(ctx context.Context, containerID string, timeout *time.Duration) error {
-	return nil
-}
-
-func (m *mockDockerClientWithoutCw) ContainerRemove(ctx context.Context, containerID string, options types.ContainerRemoveOptions) error {
-	return nil
-}
-
-func (m *mockDockerClientWithoutCw) ContainerInspect(ctx context.Context, containerID string) (types.ContainerJSON, error) {
-	return types.ContainerJSON{
-		ContainerJSONBase: &types.ContainerJSONBase{
-			HostConfig: &container.HostConfig{
-				AutoRemove: true,
-			},
-		},
-	}, nil
-}
-
-func (m *mockDockerClientWithoutCw) DistributionInspect(ctx context.Context, image, encodedRegistryAuth string) (registry.DistributionInspect, error) {
-	return registry.DistributionInspect{
-		Descriptor: v1.Descriptor{
-			Digest: "sha256:7173b809",
-		},
-	}, nil
-}
-
-func (m *mockDockerClientWithoutCw) RegistryLogin(ctx context.Context, auth types.AuthConfig) (registry.AuthenticateOKBody, error) {
-	return registry.AuthenticateOKBody{}, nil
-}
-
-// This mock client will return errors for each call to a docker function
-type mockDockerErrorClient struct {
-}
-
-var errImagePull = errors.New("error pulling image")
-var errImageList = errors.New("error listing images")
-var errContainerList = errors.New("error listing containers")
-var errContainerStop = errors.New("error stopping container")
-var errContainerRemove = errors.New("error removing container")
-var errContainerInspect = errors.New("error inspecting container")
-var errDistributionInspect = errors.New("error inspecting distribution")
-
-func (m *mockDockerErrorClient) ImageList(ctx context.Context, imageListOptions types.ImageListOptions) ([]types.ImageSummary, error) {
-	return nil, errImageList
-}
-
-func (m *mockDockerErrorClient) ImagePull(ctx context.Context, image string, imagePullOptions types.ImagePullOptions) (io.ReadCloser, error) {
-	r := ioutil.NopCloser(bytes.NewReader([]byte("")))
-	return r, errImagePull
-}
-
-func (m *mockDockerErrorClient) ContainerList(ctx context.Context, containerListOptions types.ContainerListOptions) ([]types.Container, error) {
-	return []types.Container{}, errContainerList
-}
-
-func (m *mockDockerErrorClient) ContainerStop(ctx context.Context, containerID string, timeout *time.Duration) error {
-	return errContainerStop
-}
-
-func (m *mockDockerErrorClient) ContainerRemove(ctx context.Context, containerID string, options types.ContainerRemoveOptions) error {
-	return errContainerRemove
-}
-
-func (m *mockDockerErrorClient) ContainerInspect(ctx context.Context, containerID string) (types.ContainerJSON, error) {
-	return types.ContainerJSON{}, errContainerInspect
-}
-
-func (m *mockDockerErrorClient) DistributionInspect(ctx context.Context, image, encodedRegistryAuth string) (registry.DistributionInspect, error) {
-	return registry.DistributionInspect{}, errDistributionInspect
-}
-
-func (m *mockDockerErrorClient) RegistryLogin(ctx context.Context, auth types.AuthConfig) (registry.AuthenticateOKBody, error) {
-	return registry.AuthenticateOKBody{}, nil
-}
 
 func TestPullImage(t *testing.T) {
 	t.Run("does not error when docker ImagePull succeeds", func(t *testing.T) {
